% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cnot.R
\name{Cnot}
\alias{Cnot}
\title{Cnot gate}
\usage{
Cnot(a, b)
}
\description{
cnot gate for a circuit with two qubits
}
\examples{
# to apply Cnot gate to a circuit called qc

for Cx(0,1)
Cnot(qc,01)

forcx(1,0)
Cnot(qc,10)

}
